Connected Schedule to redux
