+++ autoscheduler/frontend/src/tests/App.test.tsx
@@ -1,10 +1,14 @@
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import * as React from 'react';
+import { Provider } from 'react-redux';
+import { createStore } from 'redux';
 
 import App from '../components/App/App';
+import autoSchedulerReducer from '../redux/reducers';
 
 test('renders without errors', () => {
-  const { container } = render(<App />);
+  const store = createStore(autoSchedulerReducer);
+  const { container } = render(<Provider store={store}><App /></Provider>);
   expect(container).toBeTruthy();
 });
